fix: cannot find module 'devextreme/**' or its corresponding type declarations
Reason:

The `pack` target triggers the `build` target.

```
  "pack": {
    "executor": "nx:run-commands",
    "dependsOn": ["^build"], <==== here the build target of deps is triggered
    "options": {
      "commands": [
        "pnpm --workspace-root nx build devextreme-react", <==== here the build target is triggered, which has `"dependsOn": ["^build"]`
        "pnpm --workspace-root nx license-headers devextreme-react",
      ],
      "parallel": false
    },
```
packages/devextreme-react/project.json (same in the Vue project.json)

It was added to avoid duplication of the same sequence of commands as in the `build`  target.
Negative side-effect of that is triggering the `build` target with the `BUILD_INTERNAL_PACKAGE=true` (it cleans the `packages/devextreme/artifacts/npm/devextreme` folder),  but the `tools/scripts/build-all.ts` launches `pnpx nx build-dist devextreme` with `BUILD_INTERNAL_PACKAGE=false`.

Solution:

1) Remove the `"dependsOn": ["^build"]` from the `pack` target
2) Explicitly specify commands in the `pack` target